// src/app/models/kevaCharge.model.ts
export interface KevaCharge {
      ChargeAmountrealMatbeazar: number;
      ChargeAmountrealShekl: number;
      CloseDate: string;
      EmpName: string;
      KevaChargeId: number;
      Status: number;
      ValueDate: string;
      chargedate: string;
      countrows: number;
      employeeIdEnd: number;
      employeeIdStart: number;
      instituteCode: string;
      instituteId: number;
      instituteName: string;
      remark: string;
      
}