// src/app/models/kevaChargeById.model.ts
export interface KevaChargeById {
      Accountno: string;
      ChargeAmount: number;
      CreditCardAproveCode: string;
      CurrencyId: string;
      Customerid: number;
      FileAs: string;
      KEVANAME: string;
      KevaChargeId: number;
      Kevaid: number;
      RecieptNo: string;
      RecieptTypeId: number;
      Remark: string;
      ReturnResonId: number;
      ReturnResonname: string;
      TotalInLeadCurrent: number;
      ValueDate: string;
      chargedate: string;
      id: number;
      tadirut: number;
}